Criaçao da base do projeto
