+++ src/OpenAi/openAiConfig.ts
@@ -1,0 +1,14 @@
+import OpenAI from "openai";
+import readline from "readline";
+
+
+
+const openAiApi = new OpenAI({
+    organization: "org-rDrlShdTz7ROCkDm1gT2l7vZ",
+    apiKey: "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",
+  });
+
+  export default openAiApi
+  // Modelo de request // Atentar se para o modelo do request
+
+
